+++ drift/test/platforms/ffi/native_database_test.dart
@@ -1,8 +1,11 @@
 @TestOn('vm')
+import 'dart:ffi';
 import 'dart:io';
 
 import 'package:drift/drift.dart';
 import 'package:drift/native.dart';
+import 'package:drift/src/sqlite3/database.dart';
+import 'package:sqlite3/open.dart';
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
@@ -120,6 +123,47 @@ void main() {
     // Should also prevent subsequent open attempts
     await expectLater(db.ensureOpen(_FakeExecutorUser()), throwsA(exception));
   });
+
+  test('disposes statements directly if cache is disabled', () async {
+    final underlying = sqlite3.openInMemory();
+    final db =
+        NativeDatabase.opened(underlying, cachePreparedStatements: false);
+    addTearDown(db.close);
+
+    await db.ensureOpen(_FakeExecutorUser());
+    await db.runCustom('CREATE TABLE foo (bar INTEGER);');
+    await db.runCustom('UPDATE foo SET bar = 1');
+    expect(underlying.activeStatementCount, isZero);
+  });
+
+  test('disposes statements eventually if cache is enabled', () async {
+    final underlying = sqlite3.openInMemory();
+    addTearDown(underlying.dispose);
+    final db = NativeDatabase.opened(
+      underlying,
+      cachePreparedStatements: true,
+      closeUnderlyingOnClose: false,
+    );
+
+    await db.ensureOpen(_FakeExecutorUser());
+    expect(underlying.activeStatementCount, isZero);
+
+    for (var i = 1; i <= PreparedStatementsCache.defaultSize; i++) {
+      await db.runSelect('SELECT $i', []);
+      expect(underlying.activeStatementCount, i);
+    }
+
+    for (var i = 0; i < PreparedStatementsCache.defaultSize; i++) {
+      // This will evict old statements, so the amount of active statements
+      // should not change.
+      await db.runSelect('SELECT $i AS "new"', []);
+      expect(
+          underlying.activeStatementCount, PreparedStatementsCache.defaultSize);
+    }
+
+    await db.close();
+    expect(underlying.activeStatementCount, isZero);
+  });
 }
 
 class _FakeExecutorUser extends QueryExecutorUser {
@@ -130,4 +174,27 @@ class _FakeExecutorUser extends QueryExecutorUser {
 
   @override
   int get schemaVersion => 1;
+}
+
+extension on Database {
+  /// Counts the amount of statements currently prepared in this database by
+  /// invoking `sqlite3_next_stmt` repeatedly.
+  int get activeStatementCount {
+    final library = open.openSqlite();
+    final nextStmt = library.lookupFunction<Pointer Function(Pointer, Pointer),
+        Pointer Function(Pointer, Pointer)>(
+      'sqlite3_next_stmt',
+    );
+
+    Pointer currentStatement;
+    var count = 0;
+
+    for (currentStatement = nextStmt(handle.cast(), nullptr);
+        currentStatement.address != 0;
+        currentStatement = nextStmt(handle.cast(), currentStatement)) {
+      count++;
+    }
+
+    return count;
+  }
 }